update getLacuna in ClientRepositoryImpl

+++ app/src/main/java/com/kotleters/mobile/feature/client/domain/entity/LacunaCreate.kt
@@ -1,7 +1,7 @@
 package com.kotleters.mobile.feature.client.domain.entity
 
 data class LacunaCreate(
-    val amount: Long,
+    val amount: Double,
     val text: String,
-    val categoryId: Int,
+    val categoryId: Long,
 )
